Fix compiler warning: implicit conversion
grisbi/src/gsb_rgba.c:675:13: warning: implicit conversion loses integer precision: 'unsigned long' to 'gint' (aka 'int') [-Wshorten-64-to-32]
                        length = strlen (tab_property[1]);
                               ~ ^~~~~~~~~~~~~~~~~~~~~~~~

+++ src/gsb_rgba.c
@@ -667,7 +667,7 @@ gchar *gsb_css_get_rule_from_name (const gchar *name,
 		if (tmp_str)
 		{
 			gchar **tab_property;
-			gint length;
+			gsize length;
 
 			tab_property = g_strsplit (tmp_str, ":", 2);
